+++ Tests/Export-PsdXml.test.ps1
@@ -3,7 +3,7 @@
 
 #! on problems, the target file should not be touched
 task BadXml {
-	Remove-Item [z].psd1
+	remove z.psd1
 	$xml = [xml]'<bar><bad/></bar>'
 	($r = try {Export-PsdXml z.psd1 $xml} catch {$_})
 	equals "$r" "Unexpected node 'bad'."